--- workbench_access_protect/workbench_access_protect/workbench_access_protect.info.yml
@@ -1,7 +1,0 @@
-name: Workbench Access Protect
-type: module
-description: Adds additional logic upon entities for delete operations.
-core: 8.x
-package: Workbench
-dependencies:
-  - workbench_access:workbench_access
