--- workbench_access_protect/workbench_access_protect/workbench_access_protect.module
@@ -1,42 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains workbench_access_protect.module.
- */
-
-use Drupal\Core\Entity\EntityInterface;
-use Drupal\Core\Session\AccountInterface;
-use Drupal\Core\Access\AccessResult;
-use Drupal\workbench_access_protect\Access\Taxonomy;
-
-
-/**
- * Implements hook_entity_access().
- */
-function workbench_access_protect_entity_access(EntityInterface $entity, $op, AccountInterface $account) {
-
-  if ($op == 'delete') {
-
-    /** @var Drupal\workbench_access_protect\Access\DeleteAccessCheck $protect */
-    $protect = \Drupal::service('workbench_access_protect.delete');
-
-    if ($protect->hasBundles($entity)) {
-      $bundle = $protect->getBundles($entity)[$entity->id()];
-      // Check the entire bundle to see if any term is in use.
-      if ($protect->isDeleteAllowedBundle($bundle, $entity) === FALSE) {
-        return AccessResult::forbidden();
-      }
-    }
-    else {
-      // We're at the leaf entity and can check the specific entity.
-      if ($protect->isDeleteAllowed($entity) === FALSE) {
-          return AccessResult::forbidden();
-      }
-
-    }
-  }
-
-  return AccessResult::neutral();
-
-}
